requirements.txt
python-dotenv
openai
langchain
langchain-core
langchain-openai